2024.06.02 공지사항 & 구매 후기 게시판(일부) 부트스트랩 적용 및 파일삭제 부분 수정

+++ carcarocean/src/main/java/kr/notice/action/DeleteFileAction.java
@@ -25,23 +25,28 @@ public class DeleteFileAction implements Action{
 		
 		NoticeDao dao = NoticeDao.getDao();
 		NoticeVo db_notice = dao.getNotice(notice_num);
-		dao.deleteFile(notice_num);
 		
-		
-		//파일 삭제
-		String[] photos = db_notice.getNotice_photo().split(",");
-		for(String pho : photos) {
-			FileUtil.removeFile(request, pho);
+		//db_notice에 담긴 notice_photo 가 null 이 아닌 경우에만 실행
+		if(db_notice.getNotice_photo() != null) {
+			dao.deleteFile(notice_num);
+			
+			//파일 삭제
+			String[] photos = db_notice.getNotice_photo().split(",");
+			
+			for(String pho : photos) {
+				FileUtil.removeFile(request, pho);
+			}
 		}
 		
 		mapAjax.put("result", "success");
+		
 		//JSON 데이터 생성
 		ObjectMapper mapper = new ObjectMapper();
 		String ajaxData = mapper.writeValueAsString(mapAjax);
 		
 		request.setAttribute("ajaxData", ajaxData);
 		
-		return "/WEB-INF/views/common/ajax_view.jsp";
+		return "/WEB-INF/views/common/ajax_view.jsp";	
 	}
 	
 }
